+++ packages/core/styles/src/index.ts
@@ -33,6 +33,7 @@ export { radio } from "./packages/atomic/radio";
 export { select } from "./packages/atomic/select";
 export { skeleton } from "./packages/atomic/skeleton";
 export { spinner } from "./packages/atomic/spinner";
+export { slider } from "./packages/atomic/slider";
 export { toggle } from "./packages/atomic/toggle";
 export { link } from "./packages/atomic/link";
 export { tooltip } from "./packages/atomic/tooltip";