+++ packages/react/src/atomic/Slider/src/hooks/index.ts
@@ -1,0 +1,3 @@
+export { useSliderValues } from "./useSliderValues";
+export { useSliderDrag } from "./useSliderDrag";
+export { useSingleSliderDrag } from "./useSingleSliderDrag";