+++ packages/react/src/index.ts
@@ -23,6 +23,7 @@ export * from "@nimbus-ds/radio";
 export * from "@nimbus-ds/select";
 export * from "@nimbus-ds/skeleton";
 export * from "@nimbus-ds/spinner";
+export * from "@nimbus-ds/slider";
 export * from "@nimbus-ds/tag";
 export * from "@nimbus-ds/text";
 export * from "@nimbus-ds/textarea";